requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blis==0.7.10
CacheControl==0.13.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.6
confection==0.1.1
configobj==5.0.8
configparser==6.0.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
Django==4.2.4
django-sass==1.1.0
en-SkillExtraction @ https://huggingface.co/MohammedShaneeb/en_SkillExtraction/resolve/main/en_SkillExtraction-any-py3-none-any.whl
etelemetry==0.3.0
executing==1.2.0
filelock==3.12.2
firebase-admin==6.2.0
fitz==0.0.1.dev2
fonttools==4.42.0
frozenlist==1.4.0
future==0.18.3
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
importlib-resources==6.0.1
ipython==8.14.0
isodate==0.6.1
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
langcodes==3.3.0
libsass==0.22.0
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mysqlclient==2.2.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==1.5.3
pandasai==0.8.4
parso==0.8.3
pathlib==1.0.1
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
preshed==3.0.8
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.0
prov==2.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydot==1.4.2
Pygments==2.16.1
PyJWT==2.8.0
PyMuPDF==1.22.5
pyparsing==3.0.9
PyPDF2==3.0.1
pyshorteners==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pyxnat==1.6
rdflib==7.0.0
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.7
stack-data==0.6.2
thinc==8.1.11
threadpoolctl==3.2.0
tqdm==4.66.0
traitlets==5.9.0
traits==6.3.2
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
